Only attempt to load private key when signing URL for GCS
Attempting to create a signed url as an authorized user results in:
```
google: read JWT from JSON credentials: 'type' field is "authorized_user" (expected "service_account")
```

Indeed, only service accounts are supported because it's the only type
of credential with an embedded private key, needed for signing.

However, only the signed url operation requires it.

To allow the use of the GCS bucket's other methods when using a
non-service-account, only attempt to load the key for URLs.